:sparkles: feat: autenticação com o banco de dados e melhoria nos logs das requisições

+++ app.js
@@ -12,7 +12,7 @@ app.use(bodyParser.json());
 // Middleware logs
 app.use(function (req, res, next) {
   const agora = new DateTime(Date.now())
-  console.log(agora.toFormat("dd/MM/yyyy - HH:mm:ss"), req.method, req.url,);
+  console.log(agora.toFormat("dd/MM/yyyy - HH:mm:ss"), req.method, req.url, req.body);
   next();
 });
 
@@ -23,11 +23,16 @@ import Database from './db/Database.js';
 
 Database.init();
 const res = await Database.autenticacao()
+if(res.sucess) {
 
-// Sincronizar banco de dados e iniciar o servidor
-Database.conexao.sync({ force: true }) // Use { force: true } apenas para recriar tabelas ao desenvolver
-  .then(() => {
-    console.log('Database synced successfully.');
-    app.listen(PORT, () => console.log(`Server running on http://localhost:${PORT}`));
-  })
-  .catch(error => console.error('Error syncing database:', error));
+  // Sincronizar banco de dados e iniciar o servidor
+  Database.conexao.sync({ force: true }) // Use { force: true } apenas para recriar tabelas ao desenvolver
+    .then(() => {
+      console.log('Database synced successfully.');
+      app.listen(PORT, () => console.log(`Server running on http://localhost:${PORT}`));
+    })
+    .catch(error => console.error('Error syncing database:', error));
+
+}
+else
+  console.log(res.error);
